initial implementation of server push
The API uses loopback to the top handler in order to make sure that
the resources being pushed are related to the resource being fetched.

If the header parameter is nil, we copy the headers from the initiating
request. This is mostly a shortcut for the common case where we don't
want to specify any new request headers.