extract most logic from serverConn into conn
Most of the logic from serverConn is applicable to both client or server
conns. Separating this from the serverConn is the first step to being
able to share that code with the client/transport.

There are still some bits that are tangled up at this point. In
particular, the conn needs to call serverConn.runHandler when it has a
request ready.